# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Post Scheduler API",
        "description": "API documentation for Laravel Post Scheduler project",
        "contact": {
            "email": "raedhoussin33@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/platforms": {
            "get": {
                "tags": [
                    "Platforms"
                ],
                "summary": "Get list of all platforms",
                "operationId": "b79ea4224ff39ae923cc8756fa1bf671",
                "responses": {
                    "200": {
                        "description": "List of platforms",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Platform"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Platforms"
                ],
                "summary": "Create a new platform",
                "operationId": "057364b51fcd5d4e282c22cbc0e92eaf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Twitter"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "social"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Platform created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Platform"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/platforms/{id}": {
            "get": {
                "tags": [
                    "Platforms"
                ],
                "summary": "Get a platform by ID",
                "operationId": "3deec4f1d4d2c00065c8ec17fd57a93b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Platform ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Platform data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Platform"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Platform not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Platforms"
                ],
                "summary": "Update a platform",
                "operationId": "4d11c4ec68257f86e884985341d6f516",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Platform ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Twitter"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "social"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Platform updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Platform"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Platform not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Platforms"
                ],
                "summary": "Delete a platform",
                "operationId": "4e4d8de139b615a76288e770f777e4c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Platform ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Platform deleted"
                    },
                    "404": {
                        "description": "Platform not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/platforms/toggle": {
            "post": {
                "tags": [
                    "Platforms"
                ],
                "summary": "Enable or disable platform for authenticated user",
                "operationId": "2ed7b7d9e70d72a0b9ab0f8b61b16bfc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "platform_id",
                                    "enabled"
                                ],
                                "properties": {
                                    "platform_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "enabled": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Platform status updated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Platform": {
                "title": "Platform",
                "required": [
                    "id",
                    "name",
                    "type"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Twitter"
                    },
                    "type": {
                        "type": "string",
                        "example": "social"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-23T14:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-23T14:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token to authorize",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}